stock_price_producer/requirements.txt
pandas==2.1.0
websocket==0.2.1
websocket-client==1.6.2
avro==1.11.2
kafka-python==2.0.2